test-dev/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
pytest==6.2.4
pytest-django==4.4.0
djangorestframework==3.12.4
pytest-mock==3.6.1
drf-spectacular==0.17.2
schema==0.7.5
